+++ lib/src/bones_api_entity_adapter_postgres.dart
@@ -837,4 +837,10 @@ class PostgreSQLAdapter extends SQLAdapter<PostgreSQLExecutionContext> {
   @override
   FutureOr<void> closeTransaction(
       Transaction transaction, PostgreSQLExecutionContext? connection) {}
+
+  @override
+  String toString() {
+    var closedStr = isClosed ? ', closed' : '';
+    return 'MySQLAdapter{$databaseName@$host:$port$closedStr}';
+  }
 }